+++ NEWS.md
@@ -1,3 +1,8 @@
+PENDING: Dataiter 0.47
+======================
+
+* `DataFrame`: Fix column and method name clash errors in certain operations
+
 2023-09-05: Dataiter 0.46
 =========================